Added Channels and Drivers, Improvements to Samplers
Drivers:

The Driver object will eventually be a central place from which you can
manage all the animation for a particular 'character' offering the
ability to delay the onset of animations, blend layers (perhaps?) and
more. Currently it is a glorified dictionary

Channels:

This is the "middle layer" in the new Driver stack, the new Channel
object is responsible for managing a single quantity for example
'x_position'. It manages one or more Sampler-like objects combining them
into a single array of data which can be easily accessed.

It also supports methods for extending the data so that it cycles, or
repeats the start/end values indefintitely. It intensionally does NOT
handle time conversions and assumes the any index is a frame number,
time conversions will be left to the Driver object, when fully
implemented.

Samplers:

- Generalised Sampler to work with arbitrary (finite!) domains
- Samplers can now be added, subtracted, multiplied and negated
- Overhauled the Sampler test suite

- Also deleted the write() function as I felt this was better suited to
a plugin