Added Channels
This is the "middle layer" in the new Driver stack, the new Channel
object is responsible for managing a single quantity for example
'x_position'. It manages one or more Sampler-like objects combining them
into a single array of data which can be easily accessed.

It also supports methods for extending the data so that it cycles, or
repeats the start/end values indefintitely. It intensionally does NOT
handle time conversions and assumes the any index is a frame number,
time conversions will be left to the Driver object, when fully
implemented.