Another big refactoring
This is the first stage in introducing the concept of a `Mappable`. A
`Mappable` will be a helper class that people can use to get the ability to map
themselves onto a domain. As long as the domain fits in the standard domain
system and the mapping follows the standard pattern they will get most of this
for free.

Alas that is not what this commit is doing however, it introduces a number of
factory methods:

- `define_domain`
- `define_domain_transform`
- `define_domain_test`

Given just the name of the new domain and the parameters it has these functions
automatically write both the abstract classes needed to define the domain
interface itself and a transform on that domain. The final function will even
write the base test case for you!

Next once the concept of `Mappable` is introduced will should be able to stand
up a new color system in fairly short order.