Prototype of a helper function.
Instead of users having to deal with numpy and details like the use of
`np.logical_and` it would be good to use something like
`sty.anded(a, b, c)` and have stylo handle the details of combining the
conditions.

While the protoyped function works, I'm pretty sure the implementation
will change in response to other work.