Introduce the concept of operators (#207)
Certain definitions can now be defined as operators - that is definitions that can provide implementations of the various arithmetic operations in Python `+-*/` etc. These operators respect type annotations, meaning that addition can be defined multiple times for definitions that produce different types. 

The standard library has been updated to include operators that implement `+`, `-` and `*` for masks.