Generalise definition inputs (#216)
Definitions are longer hard coded to accept `width` and `height` as inputs, in fact it should now be possible to create definitions that accept anything as an input! The change is *mostly* transparent although

- All **positional** arguments now require type annotations.
- When evaluating definitions all inputs must now be passed as keyword arguments

As with attributes, when deriving from an existing definition any inputs on the base definition(s) will automatically inherited. Since all inputs are now explicitly typed we can ensure that definitions don't define inputs that conflict with any implicit inputs inherited from their bases.

Closes #206