Ensure compatibility with PEP 563 (#267)
As outlined in #252 having type annotations be strings at runtime by default in 3.9 would break the way we currently construct definitions. This PR should now ensure that we are future compatible with 3.9 by opting into this behaviour through a `__future__`
import. This does mean however we're dropping support for Python 3.6

Main highlights include
- Using `typing.get_type_hints()` to get the real type annotations at runtime
- Refactored the overall construction of definitions to be more explicit what is an attribute, base, input etc
- Implement custom generic type annotations the "right" way by basing `Defn` on `Generic[T]`. Hopefully this means we play nicer with the overall type hinting space 

Closes #252